+++ src/main/java/com/dnd/gongmuin/security/oauth2/Oauth2Response.java
@@ -1,4 +1,4 @@
-package com.dnd.gongmuin.security.dto;
+package com.dnd.gongmuin.security.oauth2;
 
 public interface Oauth2Response {
 	String getProvider();
